% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download_ILDIS}
\alias{download_ILDIS}
\title{Download ILDIS htnl file, with legume species information}
\usage{
download_ILDIS(species, hdir = "ILDIS_html")
}
\arguments{
\item{species}{Latin binomial name for legume species of interest}

\item{hdir}{Directory to store the downloaded html file}
}
\description{
Download ILDIS htnl file, with legume species information
}

