% Generated by roxygen2 (4.0.0): do not edit by hand
\name{extract_ILDIS}
\alias{extract_ILDIS}
\title{Extract geographical records data from ILDIS species page html}
\usage{
extract_ILDIS(html_file)
}
\arguments{
\item{html_file}{The path of the html file to process}
}
\description{
Extract geographical records data from ILDIS species page html
}
\details{
The name of the species is extracted from the filename, as generated by \code{\link{download_ILDIS}}.
}

