% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{legume_species}
\alias{legume_species}
\title{List of Legume Species in Werner et al. Dataset}
\usage{
data(legume_species)
}
\description{
This is a vector of the species found in the nitrogen fixation data of Werner et al (2014).
Mostly used to get data from other databases.
}

